requirements.txt
aiohttp==3.8.1
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@fbb8803779373357e274e1540b368365fd9d8074
discord.py @ git+https://github.com/Rapptz/discord.py@5a06fa5f3e28d2b7191722e1a84c541560008aea
jishaku==2.5.0
python-dotenv==0.20.0
